ci: cache nox
Right now, we recreate the nox environment on every CI run, which is wasteful
as it can take a while to to create it.

This patch caches the nox environment so that it doesn't need to be
recreated on every CI run. The cache will only be updated once a day.

If at some point we save a bad environment or the cached environment
becomes incompatible with newer nox versions, we should bump the
version (currently v1) in the cache key, to force the cache to be
recreated.

Signed-off-by: Filipe Laíns <lains@archlinux.org>