Provide new copy_file implementation
This patch adds a new IOHandler class which implements the "writer"
interface under the `copy_file` method to perform the copy operation in
installer. There is no formal interface defined yet, but the expectation
is that it will be a single function which performs the copy from source
to destination and returns a Record type containing the path, hash and a
size values.