Change `parse_record_file` to yield the elements
This is necessary to ensure that the conversion from elements
to `RecordEntry` is done only once. Since we're passing the elements
across the `WheelSource` API, the call to `RecordEntry.from_elements`
is explicitly done in the `install` function. This renders this parsing
function useless for use within the `WheelSource` to parse RECORD files.

By decoupling the creation on `RecordEntry` objects from these elements,
this change enables `WheelSource` implementations to use this to parse
the `RECORD` file.