installer: add --interpreter option
The currently running Python might not be the right one for the installed
wheels, so add an option to allow the interpreter to be overridden.

This is essential in cross/distro builds where we want to install a
wheel into a working directory to build a distro package (eg a rpm or
deb) with pypa/installer, but the currently running python on the build
host might not be the right path for the scripts to use on the target.