Encode hashes in RECORD files correctly according to PEP 376
Installer was generating dist-info RECORD files with invalid hashes.
Specifically, it was encoding the hashes in hexadecimal, but PEP 376
specifies that they are to be encoded in urlsafe-base64-nopad.
Bizarrely, Installer's own records.Hash class implements a validate
method that checks a recorded hash against the correctly encoded hash of
the corresponding file, meaning that Installer could never validate the
very records that it itself had generated.

Fix the utils.copyfileobj_with_hashing function so that it returns
correctly encoded hashes, and fix up TestCopyFileObjWithHashing and
TestSchemeDictionaryDestination to check that the hashes are encoded as
specified by PEP 376.