Normalize `RECORD` file paths when parsing
The `RECORD` file is permitted contain paths with backslashes (eg: on
Windows). Normalising these paths to forward slashes on parse makes it
so that consumers of the `RECORD` file that use `parse_record_file`
do not need to worry about handling this edge case/normalisation. This
does mean that `RECORD` parsing with this function would not be
round-trip safe, but that is not an intended use case.

Co-Authored-By: Pradyun Gedam <pradyunsg@gmail.com>