Treat an empty destination the same as an unspecified one
With a command line such as:
```
python -m installer --destdir="$DESTDIR" dist/*.whl
```

The destdir may not be "set" but is always passed. When relocating paths
relative to the destination directory, the anchor is chopped off,
resulting in paths that are relative to the process-wide current working
directory, and are installed relative to there. Avoid doing any
relocations against an empty destdir.

Fixes #136